move component to proper location
